+++ lib/cogment/services/actor_service.cpp
@@ -17,6 +17,8 @@
 #include "cogment/orchestrator.h"
 #include "cogment/versions.h"
 
+#include "spdlog/spdlog.h"
+
 #include <string>
 
 namespace cogment {
@@ -29,27 +31,60 @@ ActorService::ActorService(Orchestrator* orch) : orchestrator_(orch) {}
 ::easy_grpc::Stream_future<::cogment::TrialActionReply> ActorService::ActionStream(
     ::easy_grpc::Stream_future<::cogment::TrialActionRequest> actions, easy_grpc::Context ctx) {
   auto trial_id = uuids::uuid::from_string(ctx.get_client_header("trial-id"));
-  auto actor_name = std::string(ctx.get_client_header("actor-name"));
+  std::string actor_name(ctx.get_client_header("actor-name"));
 
   return orchestrator_->bind_client(trial_id, actor_name, std::move(actions));
 }
 
 ::easy_grpc::Future<::cogment::TrialHeartbeatReply> ActorService::Heartbeat(::cogment::TrialHeartbeatRequest,
-                                                                            easy_grpc::Context ctx) {
-  (void)ctx;
+                                                                            easy_grpc::Context) {
   return {};
 }
 
-::easy_grpc::Future<::cogment::TrialRewardReply> ActorService::SendReward(::cogment::TrialRewardRequest,
+::easy_grpc::Future<::cogment::TrialRewardReply> ActorService::SendReward(::cogment::TrialRewardRequest reward,
                                                                           easy_grpc::Context ctx) {
-  (void)ctx;
-  return {};
+  auto trial_id = uuids::uuid::from_string(ctx.get_client_header("trial-id"));
+  std::string actor_name(ctx.get_client_header("actor-name"));
+
+  try {
+    auto trial = orchestrator_->get_trial(trial_id);
+
+    for (auto& rew : reward.rewards()) {
+      trial->reward_received(rew, actor_name);
+    }
+  } catch (const std::out_of_range& oor) {
+    spdlog::error("Trial {} doesn't exit: {}", to_string(trial_id), oor.what());
+  }
+
+  ::easy_grpc::Promise<::cogment::TrialRewardReply> prom;
+  auto result = prom.get_future();
+  ::cogment::TrialRewardReply reply;
+  prom.set_value(std::move(reply));
+
+  return result;
 }
 
-::easy_grpc::Future<::cogment::TrialMessageReply> ActorService::SendMessage(::cogment::TrialMessageRequest,
+::easy_grpc::Future<::cogment::TrialMessageReply> ActorService::SendMessage(::cogment::TrialMessageRequest message,
                                                                             easy_grpc::Context ctx) {
-  (void)ctx;
-  return {};
+  auto trial_id = uuids::uuid::from_string(ctx.get_client_header("trial-id"));
+  std::string actor_name(ctx.get_client_header("actor-name"));
+
+  try {
+    auto trial = orchestrator_->get_trial(trial_id);
+
+    for (auto& mess : message.messages()) {
+      trial->message_received(mess, actor_name);
+    }
+  } catch (const std::out_of_range& oor) {
+    spdlog::error("Trial {} doesn't exit: {}", to_string(trial_id), oor.what());
+  }
+
+  ::easy_grpc::Promise<::cogment::TrialMessageReply> prom;
+  auto result = prom.get_future();
+  ::cogment::TrialMessageReply reply;
+  prom.set_value(std::move(reply));
+
+  return result;
 }
 
 ::cogment::VersionInfo ActorService::Version(::cogment::VersionRequest, easy_grpc::Context ctx) {