+++ CHANGELOG.md
@@ -8,10 +8,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 
 ### Added
-
+- Add support for messages and rewrads, they can be sent from the actor client.
 - Addition for tls communication
 
 ## v1.0.0-alpha4 - 2021-02-17
+
+### Added
+- Add support for messages and rewrads, they can be sent from the actor client.
 
 ## v1.0.0-alpha3 - 2021-01-28